Move search form into Spaces component
