+++ rehearsal-room-frontend/src/components/Application.jsx
@@ -15,7 +15,7 @@ import HeroV1 from './HeroBlock/heroV1'
 import Space from './Space'
 
 import Spaces from './Spaces'
-import SearchForm from './Spaces/SearchForm'
+import SearchForm from './SearchForm'
 
 // Images
 import logo from '../../src/logo.svg';
@@ -52,7 +52,6 @@ export default function App() {
         <Switch>
           <Route path="/spaces/:city">
             <HeroV1 />
-            <SearchForm />
             <Spaces />
           </Route>
           <Route path="/space/:space_id">
